+++ private/build_schema/build_mutation_fields/configuration.js
@@ -5,7 +5,7 @@ const defaultValue = {
   blocksBehind: 3,
   expireSeconds: 30,
   max_net_usage_words: 0,
-  max_cpu_usage_ms: 1,
+  max_cpu_usage_ms: 0,
   delay_sec: 0
 }