+++ private/serialize/extension.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const serializeUint = require('./uint.js')
-const serializeVaruint32 = require('./varuint32.js')
+const serializeUint = require('eosio-wasm-js/public/serialize/uint.js')
+const serializeVaruint32 = require('eosio-wasm-js/public/serialize/varuint32.js')
 
 const serialize_extension = ({ type, data }) =>
   serializeUint(type) + serializeVaruint32(data.length / 2) + data